+++ packages/block-editor/src/components/block-list/use-block-props/index.js
@@ -1,20 +1,9 @@
-/**
- * External dependencies
- */
 import clsx from 'clsx';
-
-/**
- * WordPress dependencies
- */
 import { useContext } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 import { __unstableGetBlockProps as getBlockProps } from '@wordpress/blocks';
 import { useMergeRefs, useDisabled, useRefEffect } from '@wordpress/compose';
 import warning from '@wordpress/warning';
-
-/**
- * Internal dependencies
- */
 import useMovingAnimation from '../../use-moving-animation';
 import { PrivateBlockContext } from '../private-block-context';
 import { useFocusFirstElement } from './use-focus-first-element';
@@ -107,6 +96,7 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 		blockVisibility,
 		deviceType,
 		viewportSettings,
+		ariaLabel,
 	} = useContext( PrivateBlockContext );
 
 	useRegisterBlockEventHandlers( clientId, wrapperProps );
@@ -186,7 +176,7 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 		ref: mergedRefs,
 		id: `block-${ clientId }${ htmlSuffix }`,
 		role: 'document',
-		'aria-label': blockLabel,
+		'aria-label': ariaLabel ?? blockLabel,
 		'data-block': clientId,
 		'data-type': name,
 		'data-title': blockTitle,