+++ packages/editor/src/components/post-actions/set-as-posts-page.js
@@ -1,6 +1,3 @@
-/**
- * WordPress dependencies
- */
 import { __, sprintf } from '@wordpress/i18n';
 import { useMemo } from '@wordpress/element';
 import {
@@ -12,10 +9,6 @@ import {
 import { useDispatch, useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
 import { store as noticesStore } from '@wordpress/notices';
-
-/**
- * Internal dependencies
- */
 import { getItemTitle } from '../../utils/get-item-title';
 
 const SetAsPostsPageModal = ( { items, closeModal } ) => {