+++ packages/fields/src/fields/parent/parent-edit.tsx
@@ -1,11 +1,4 @@
-/**
- * External dependencies
- */
 import removeAccents from 'remove-accents';
-
-/**
- * WordPress dependencies
- */
 import { ComboboxControl, ExternalLink } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
 import {
@@ -14,17 +7,12 @@ import {
 	useMemo,
 	useState,
 } from '@wordpress/element';
-// @ts-ignore
 import { store as coreStore } from '@wordpress/core-data';
 import type { DataFormControlProps } from '@wordpress/dataviews';
 import { debounce } from '@wordpress/compose';
 import { decodeEntities } from '@wordpress/html-entities';
 import { __, sprintf } from '@wordpress/i18n';
 import { filterURLForDisplay } from '@wordpress/url';
-
-/**
- * Internal dependencies
- */
 import type { BasePost } from '../../types';
 import { getTitleWithFallbackName } from './utils';