+++ packages/rich-text/src/test/insert.js
@@ -1,12 +1,4 @@
-/**
- * External dependencies
- */
 import deepFreeze from 'deep-freeze';
-
-/**
- * Internal dependencies
- */
-
 import { insert } from '../insert';
 import { getSparseArrayLength } from './helpers';